refactor_project_structure(structure): refactor project
