+++ trello-web/src/components/ModeSelect/index.jsx
@@ -1,0 +1,52 @@
+import { useColorScheme } from '@mui/material'
+import InputLabel from '@mui/material/InputLabel'
+import MenuItem from '@mui/material/MenuItem'
+import FormControl from '@mui/material/FormControl'
+import Select from '@mui/material/Select'
+import Box from '@mui/material/Box'
+
+import LightModeIcon from '@mui/icons-material/LightMode'
+import DarkModeOutlinedIcon from '@mui/icons-material/DarkModeOutlined'
+import SettingBrightnessIcon from '@mui/icons-material/SettingsBrightness'
+
+const ModeSelect = () => {
+  const { mode, setMode } = useColorScheme()
+
+  const handleChange = (event) => {
+    setMode(event.target.value)
+  }
+
+  return (
+    <FormControl sx={{ m: 1, minWidth: 120 }} size="small">
+      <InputLabel id="label-select-dark-light-mode">Mode</InputLabel>
+      <Select
+        labelId="label-select-dark-light-mode"
+        id="select-dark-light-mode"
+        value={mode}
+        label="Mode"
+        onChange={handleChange}
+      >
+        <MenuItem value="light">
+          <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+            <LightModeIcon fontSize="small" />
+            Light
+          </Box>
+        </MenuItem>
+        <MenuItem value="dark">
+          <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+            <DarkModeOutlinedIcon fontSize="small" />
+            Dark
+          </Box>
+        </MenuItem>
+        <MenuItem value="system">
+          <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+            <SettingBrightnessIcon fontSize="small" />
+            System
+          </Box>
+        </MenuItem>
+      </Select>
+    </FormControl>
+  )
+}
+
+export default ModeSelect
