+++ trello-web/src/pages/Auth/index.jsx
@@ -1,0 +1,1 @@
+// Auth page
