+++ trello-web/src/pages/Boards/BoardContent/index.jsx
@@ -1,0 +1,14 @@
+import Box from '@mui/material/Box'
+
+const BoardContent = () => {
+  return (
+    <Box sx={{
+      backgroundColor: 'primary.main',
+      width: '100%',
+      height: (theme) => `calc(100vh - ${theme.trello.appBarHeight} - ${theme.trello.boardBarHeight})`,
+      display: 'flex',
+      alignItems: 'center'
+    }}>Content</Box>
+  )
+}
+export default BoardContent
