+++ trello-web/src/redux/store.js
@@ -1,0 +1,1 @@
+// Redux store configuration
